requirements.txt
uvicorn==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.4
psycopg2-binary==2.9.10
passlib==1.7.4
fastapi==0.115.6
python-multipart==0.0.20
python-dotenv==1.0.1
bcrypt==4.0.1
pyjwt==2.10.1